My First Mods
#osloginscreen v0.W1a

Modded to adhere to the new database standards in .9.
Modded to correct the number of users vs HG users and Total number.
Modded to ONLY show local regions in the region list, ignoring HG linked
Regions.
Modded theme to add a touch of transparency. (this can be seen here
http://opensim.life/osloginscreen/)

Removed the "fork me on github" corrner image, just for usability sake.